+++ vfs/vfs.c
@@ -1300,7 +1300,7 @@ int vfs_unlink(const char* path) {
   return error;
 }
 
-int vfs_read(int fd, void* buf, size_t count) {
+ssize_t vfs_read(int fd, void* buf, size_t count) {
   file_t* file = 0x0;
   int result = lookup_fd(fd, &file);
   if (result) return result;
@@ -1348,7 +1348,7 @@ int vfs_read(int fd, void* buf, size_t count) {
   return result;
 }
 
-int vfs_write(int fd, const void* buf, size_t count) {
+ssize_t vfs_write(int fd, const void* buf, size_t count) {
   file_t* file = 0x0;
   int result = lookup_fd(fd, &file);
   if (result) return result;
